split json_schema.rb to smaller files
