+++ lib/rsmp.rb
@@ -107,5 +107,4 @@ require_relative 'rsmp/tlc/signal_group'
 require_relative 'rsmp/tlc/signal_plan'
 require_relative 'rsmp/tlc/input_states'
 require_relative 'rsmp/tlc/signal_priority'
-require_relative 'rsmp/convert/import/yaml'
 require_relative 'rsmp/version'
